+++ CHANGELOG.md
@@ -960,3 +960,4 @@
 - [2026-05-06 19:54] fix(security): prevent double tap duplicate action in security
 - [2026-05-06 20:24] feat(analytics): add async analytics stream provider
 - [2026-05-06 20:49] feat(cache): add bi-directional event bus handler for cache
+- [2026-05-06 22:04] chore(analytics): update dependencies and cleanup unused imports in analytics